+++ R/gfcmSTAR-package.R
@@ -23,7 +23,7 @@
 #' }
 #' \emph{Export:}
 #' \tabular{ll}{
-#'   \code{\link{write.star}} \tab write STAR object to CSV files
+#'   \code{\link{write.star.csv}} \tab write STAR object to CSV files
 #' }
 #' \emph{Format strings:}
 #' \tabular{ll}{