+++ R/read.star.csv.R
@@ -15,6 +15,8 @@
 #'
 #' \code{\link{read.template}} reads an Excel STAR template into a STAR object.
 #'
+#' \code{\link{write.star.csv}} writes a STAR object to CSV files.
+#'
 #' \code{\link{gfcmSTAR-package}} gives an overview of the package.
 #'
 #' @examples