+++ R/read.star.csv.R
@@ -10,8 +10,8 @@
 #' \code{TimeSeries} (data frame).
 #'
 #' @seealso
-#' \code{\link{read.csv}} is the underlying base function used to read the CSV
-#' files.
+#' \code{\link{import.many.csv}} imports many STAR CSV files from a directory
+#' tree into a cluster (list).
 #'
 #' \code{\link{read.template}} reads an Excel STAR template into a STAR object.
 #'